+++ CHANGELOG.md
@@ -9,6 +9,7 @@ HEAD
 * Change mqtt_client_event_data.data_len to size_t
 * Fix string debug message format
 * Remove conflicting defines for esp8266
+* Remove WebConfigServer from basic_class_with_mqtt and update example with latest version
 
 v0.0.8 (2023-05-09)
 ------